+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* Basic seeding for Participants, linking Projects to random People
 * Project acronym property
 * Project abstract property
 * Project publications property